+++ spec/builder/exact_path_spec.rb
@@ -5,7 +5,7 @@ RSpec.describe PathList::Builder::ExactPath do
 
   let(:path) { '/path/to/exact/something' }
   let(:candidate_path) { path }
-  let(:candidate) { PathList::Candidate.new(candidate_path, true, nil) }
+  let(:candidate) { PathList::Candidate.new(candidate_path, true) }
   let(:polarity) { :allow }
 
   describe '#build_implicit' do
@@ -38,50 +38,50 @@ RSpec.describe PathList::Builder::ExactPath do
     end
 
     it "doesn't need to match exact path" do
-      expect(matcher.match(PathList::Candidate.new(path, true, nil))).to be_nil
+      expect(matcher.match(PathList::Candidate.new(path, true))).to be_nil
     end
 
     it 'matches most parent path' do
-      expect(matcher.match(PathList::Candidate.new('/path', true, nil)))
+      expect(matcher.match(PathList::Candidate.new('/path', true)))
         .to be :allow
     end
 
     it 'matches most parent path regardless of case' do
-      expect(matcher.match(PathList::Candidate.new('/PATH', true, nil))).to be :allow
+      expect(matcher.match(PathList::Candidate.new('/PATH', true))).to be :allow
     end
 
     it 'matches parent path' do
-      expect(matcher.match(PathList::Candidate.new('/path/to', true, nil)))
+      expect(matcher.match(PathList::Candidate.new('/path/to', true)))
         .to be :allow
     end
 
     it 'matches child path' do
-      expect(matcher.match(PathList::Candidate.new("#{path}/child", true, nil)))
+      expect(matcher.match(PathList::Candidate.new("#{path}/child", true)))
         .to be :allow
     end
 
     it 'matches grandchild path' do
-      expect(matcher.match(PathList::Candidate.new("#{path}/child/child", true, nil)))
+      expect(matcher.match(PathList::Candidate.new("#{path}/child/child", true)))
         .to be :allow
     end
 
     it "doesn't match path starting with the same string" do
-      expect(matcher.match(PathList::Candidate.new("#{path}_part_2", true, nil)))
+      expect(matcher.match(PathList::Candidate.new("#{path}_part_2", true)))
         .to be_nil
     end
 
     it "doesn't match parent path starting with the same string" do
-      expect(matcher.match(PathList::Candidate.new('/path/to/exact-ish/something', true, nil)))
+      expect(matcher.match(PathList::Candidate.new('/path/to/exact-ish/something', true)))
         .to be_nil
     end
 
     it "doesn't match path sibling" do
-      expect(matcher.match(PathList::Candidate.new('/path/to/exact/other', true, nil)))
+      expect(matcher.match(PathList::Candidate.new('/path/to/exact/other', true)))
         .to be_nil
     end
 
     it "doesn't match path concatenation" do
-      expect(matcher.match(PathList::Candidate.new('/pathtoexactsomething', true, nil))) # spellr:disable-line
+      expect(matcher.match(PathList::Candidate.new('/pathtoexactsomething', true))) # spellr:disable-line
         .to be_nil
     end
   end
@@ -98,25 +98,25 @@ RSpec.describe PathList::Builder::ExactPath do
     end
 
     it 'matches exact path' do
-      expect(matcher.match(PathList::Candidate.new(path, true, nil))).to be :ignore
+      expect(matcher.match(PathList::Candidate.new(path, true))).to be :ignore
     end
 
     it 'matches exact path case insensitively' do
-      expect(matcher.match(PathList::Candidate.new(path.upcase, true, nil))).to be :ignore
+      expect(matcher.match(PathList::Candidate.new(path.upcase, true))).to be :ignore
     end
 
     it "doesn't match most parent path" do
-      expect(matcher.match(PathList::Candidate.new('/path', true, nil)))
+      expect(matcher.match(PathList::Candidate.new('/path', true)))
         .to be_nil
     end
 
     it "doesn't match parent path" do
-      expect(matcher.match(PathList::Candidate.new('/path/to', true, nil)))
+      expect(matcher.match(PathList::Candidate.new('/path/to', true)))
         .to be_nil
     end
 
     it "doesn't match child path" do
-      expect(matcher.match(PathList::Candidate.new("#{path}/child", true, nil)))
+      expect(matcher.match(PathList::Candidate.new("#{path}/child", true)))
         .to be_nil
     end
   end